+++ src/Type/InputMediaLocation.php
@@ -11,7 +11,7 @@ use Phptg\BotApi\FileCollector;
  *
  * @api
  */
-final readonly class InputMediaLocation implements InputMedia
+final readonly class InputMediaLocation implements InputPollMedia, InputPollOptionMedia
 {
     public function __construct(
         public float $latitude,