+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 - New #199: Add `PollMedia` and `LivePhoto` type.
 - New #199: Add `explanationMedia`, `media`, `membersOnly` and `countryCodes` fields to `Poll` type.
 - New #199: Add `media` field to `PollOption` type.
+- New #199: Add `InputPollMedia` and `InputPollOptionMedia` types.
 
 ## 0.20 April 26, 2026